Refactor resource creation of API objects
This changelist constructs a list of API objects to create, and attempts their creation in order, logging an optional message at the start of each `ResourceSet`’s creation. If any API object fails creation, it rolls back all created resources.

Rather than using a specific API group client for each API object, this changelist defers creation of the API client by constructing it dynamically based on the API group of the object. It then caches the API client by `GroupVersionKind`. This also refactors the code for rolling back or deleting API objects upon encountering a failure by making rollback part of the `ResourceHelper` that automatically deletes API objects if any Object in the specified set of Objects fails creation. The `ResourceHelper` also supports updating objects if they already exist, making it easier to support a `—force` or `—overwrite` flag that allows making `cilium install` idempotent.

Unlike the existing code, this approach does not require maintaining a separate list of API objects that have been created. It introduces a new type `ResourceHelper` that encapsulates creation and rolling back of API objects. It's safe for concurrent use. The code for rolling back created resources was repetitive and required repeating the namespace and name of the object to delete. The new pattern only requires specifying an API object and handles creation and rolling back of resources if required. It avoids repeating the namespace in API calls and infers the namespace from the API Object itself. When `helm.GetManifests` is called, it already sets the specified namespace on API objects, so instead of repeating the namespace obtained via CLI params, the code grabs the namespace from the API object itself.

The `ResourceHelper` is backed by tests that cover creation of resources in order, deletion of all created resources when an API object fails to create, and update of already-existing API objects.